requirements.txt
requests==2.32.3
pandas==2.2.2
python-dateutil==2.9.0.post0
matplotlib==3.9.2

